Bugfix: make sure there is only on ClassType per class
We have to collect all ClassTypes before we decent into
fields/methods. Previously it was possible that a new ClassType was
generated for a field's/method's definig class, if it was not already
present in the classTypes collection. However, in the main for loop in
visitProgram, another ClassType would be created for this class and
put into the classTypes collection, overwriting the previously created
one.